Refactor tests: Split feature tests, update project structure
