+++ requirements.txt
@@ -1,7 +1,2 @@
 # No external dependencies required for runtime
-# Uses only Python standard library
-
-# Development dependencies (for linting and formatting)
-flake8>=6.0.0
-black>=23.0.0
-isort>=5.12.0
+# This project uses only Python standard library modules
