+++ tests/feature_tests/__init__.py
@@ -1,0 +1,6 @@
+"""
+Feature tests for C to PlantUML Converter
+
+This package contains end-to-end feature tests that verify the complete
+functionality of the C to PlantUML converter tool.
+"""
